Add option to open DB in secondary mode
readonly mode

disabling custom block_cache in secondary rocksDb

fix min_masterchain_seqno

fix opening package in readonly/secondary mode

Add method RootDb::get_block_state_root

skip persistent states check in secondary/read-only modes

fix crash

fix crash when new archive slice is created

catch up all archive slices

Revert "catch up all archive slices"

This reverts commit 34c782b11e0a93db439a8033123cf8929e7ad169.

fix rebase artifact

rocksdb related fixes

specify working directory for secondary instance

Catch up prev archive slice when new one appears

Catch up archive slice before each query if more than 1 sec passed

do not crash if opening package failed in secondary mode

resolve cyclic dependency by using object lib vs static

CellDb::load_cell workaround

try catch up before getting new mc seqno

fix stuck

fix crash

+++ validator/db/archive-slice.cpp
@@ -19,6 +19,8 @@
 #include "common/delay.h"
 #include "td/actor/MultiPromise.h"
 #include "td/db/RocksDb.h"
+#include "td/db/RocksDbSecondary.h"
+#include "td/db/RocksDbReadOnly.h"
 #include "td/utils/port/path.h"
 #include "validator/fabric.h"
 
@@ -573,6 +575,14 @@ void ArchiveSlice::get_slice(td::uint64 archive_id, td::uint64 offset, td::uint3
   td::actor::create_actor<db::ReadFile>("readfile", p->path, offset, limit, 0, std::move(promise)).release();
 }
 
+void ArchiveSlice::get_max_masterchain_seqno(td::Promise<BlockSeqno> promise) {
+  promise.set_result(max_masterchain_seqno());
+}
+
+void ArchiveSlice::get_min_masterchain_seqno(td::Promise<BlockSeqno> promise) {
+  promise.set_result(min_masterchain_seqno());
+}
+
 void ArchiveSlice::get_archive_id(BlockSeqno masterchain_seqno, ShardIdFull shard_prefix,
                                   td::Promise<td::uint64> promise) {
   before_query();
@@ -593,7 +603,23 @@ void ArchiveSlice::before_query() {
     LOG(DEBUG) << "Opening archive slice " << db_path_;
     td::RocksDbOptions db_options;
     db_options.statistics = statistics_.rocksdb_statistics;
-    kv_ = std::make_unique<td::RocksDb>(td::RocksDb::open(db_path_, std::move(db_options)).move_as_ok());
+    switch (mode_) {
+      case td::DbOpenMode::db_primary:
+        kv_ = std::make_unique<td::RocksDb>(td::RocksDb::open(db_path_, std::move(db_options)).move_as_ok());
+        break;
+      case td::DbOpenMode::db_readonly:
+        kv_ = std::make_unique<td::RocksDbReadOnly>(td::RocksDbReadOnly::open(db_path_, std::move(db_options)).move_as_ok());
+        break;
+      case td::DbOpenMode::db_secondary: {
+        CHECK(secondary_workdir_);
+        td::RocksDbSecondaryOptions secondary_db_options{std::move(db_options), secondary_workdir_.value()};
+        kv_ = std::make_unique<td::RocksDbSecondary>(td::RocksDbSecondary::open(db_path_, std::move(secondary_db_options)).move_as_ok());
+        last_catch_up_ = td::Timestamp::now();
+        break;
+      }
+      default:
+        UNREACHABLE();
+    }
     std::string value;
     auto R2 = kv_->get("status", value);
     R2.ensure();
@@ -677,6 +703,11 @@ void ArchiveSlice::before_query() {
     td::actor::send_closure(archive_lru_, &ArchiveLru::on_query, actor_id(this), p_id_,
                             packages_.size() + ESTIMATED_DB_OPEN_FILES);
   }
+  if (mode_ == td::DbOpenMode::db_secondary) {
+    if (td::Timestamp::now().at() - last_catch_up_.at() > 1.0) {
+      try_catch_up_with_primary_impl().ensure();
+    }
+  }
 }
 
 void ArchiveSlice::open_files() {
@@ -743,6 +774,68 @@ void ArchiveSlice::end_async_query() {
   }
 }
 
+td::Status ArchiveSlice::try_catch_up_with_primary() {
+  CHECK(mode_ == td::DbOpenMode::db_secondary);
+  if (status_ == st_closed) {
+    before_query();
+    return td::Status::OK();
+  } else {
+    return try_catch_up_with_primary_impl();
+  }
+}
+
+td::Status ArchiveSlice::try_catch_up_with_primary_impl() {
+  CHECK(mode_ == td::DbOpenMode::db_secondary);
+  
+  TRY_STATUS(static_cast<td::RocksDbSecondary *>(kv_.get())->try_catch_up_with_primary());
+
+  std::string value;
+  auto R2 = kv_->get("status", value);
+  R2.ensure();
+  if (R2.move_as_ok() == td::KeyValue::GetStatus::Ok) {
+    if (value == "sliced") {
+      R2 = kv_->get("slices", value);
+      R2.ensure();
+      auto tot = td::to_integer<td::uint32>(value);
+      if (tot == packages_.size()) {
+        last_catch_up_ = td::Timestamp::now();
+        return td::Status::OK();
+      }
+      R2 = kv_->get("slice_size", value);
+      R2.ensure();
+      slice_size_ = td::to_integer<td::uint32>(value);
+      CHECK(slice_size_ > 0);
+      for (td::uint64 i = packages_.size(); i < tot; i++) {
+        R2 = kv_->get(PSTRING() << "status." << i, value);
+        R2.ensure();
+        auto len = td::to_integer<td::uint64>(value);
+        R2 = kv_->get(PSTRING() << "version." << i, value);
+        R2.ensure();
+        td::uint32 ver = 0;
+        if (R2.move_as_ok() == td::KeyValue::GetStatus::Ok) {
+          ver = td::to_integer<td::uint32>(value);
+        }
+        td::uint32 seqno;
+        ShardIdFull shard_prefix;
+        if (shard_split_depth_ == 0) {
+          seqno = archive_id_ + slice_size_ * i;
+          shard_prefix = ShardIdFull{masterchainId};
+        } else {
+          R2 = kv_->get(PSTRING() << "info." << i, value);
+          R2.ensure();
+          CHECK(R2.move_as_ok() == td::KeyValue::GetStatus::Ok);
+          unsigned long long shard;
+          CHECK(sscanf(value.c_str(), "%u.%d:%016llx", &seqno, &shard_prefix.workchain, &shard) == 3);
+          shard_prefix.shard = shard;
+        }
+        add_package(seqno, shard_prefix, len, ver);
+      }
+    }
+  }
+  last_catch_up_ = td::Timestamp::now();
+  return td::Status::OK();
+}
+
 void ArchiveSlice::begin_transaction() {
   if (!async_mode_ || !huge_transaction_started_) {
     kv_->begin_transaction().ensure();
@@ -781,7 +874,8 @@ void ArchiveSlice::set_async_mode(bool mode, td::Promise<td::Unit> promise) {
 
 ArchiveSlice::ArchiveSlice(td::uint32 archive_id, bool key_blocks_only, bool temp, bool finalized,
                            td::uint32 shard_split_depth, std::string db_root,
-                           td::actor::ActorId<ArchiveLru> archive_lru, DbStatistics statistics)
+                           td::actor::ActorId<ArchiveLru> archive_lru, DbStatistics statistics,
+                           td::DbOpenMode mode, td::optional<std::string> secondary_workdir)
     : archive_id_(archive_id)
     , key_blocks_only_(key_blocks_only)
     , temp_(temp)
@@ -790,7 +884,9 @@ ArchiveSlice::ArchiveSlice(td::uint32 archive_id, bool key_blocks_only, bool tem
     , shard_split_depth_(temp || key_blocks_only ? 0 : shard_split_depth)
     , db_root_(std::move(db_root))
     , archive_lru_(std::move(archive_lru))
-    , statistics_(statistics) {
+    , statistics_(statistics)
+    , mode_(mode)
+    , secondary_workdir_(std::move(secondary_workdir)) {
   db_path_ = PSTRING() << db_root_ << p_id_.path() << p_id_.name() << ".index";
 }
 
@@ -840,9 +936,13 @@ void ArchiveSlice::add_package(td::uint32 seqno, ShardIdFull shard_prefix, td::u
   } else {
     path = PSTRING() << db_root_ << p_id.path() << get_package_file_name(p_id, shard_prefix);
   }
-  auto R = Package::open(path, false, true);
+  auto R = Package::open(path, mode_ != td::DbOpenMode::db_primary, mode_ == td::DbOpenMode::db_primary);
   if (R.is_error()) {
-    LOG(FATAL) << "failed to open/create archive '" << path << "': " << R.move_as_error();
+    if (mode_ == td::DbOpenMode::db_secondary) {
+      LOG(ERROR) << "failed to open/create archive '" << path << "': " << R.move_as_error();
+    } else {
+      LOG(FATAL) << "failed to open/create archive '" << path << "': " << R.move_as_error();
+    }
     return;
   }
   if (statistics_.pack_statistics) {
@@ -855,7 +955,7 @@ void ArchiveSlice::add_package(td::uint32 seqno, ShardIdFull shard_prefix, td::u
     return;
   }
   auto pack = std::make_shared<Package>(R.move_as_ok());
-  if (version >= 1) {
+  if (version >= 1 && mode_ == td::DbOpenMode::db_primary) {
     pack->truncate(size).ensure();
   }
   auto writer = td::actor::create_actor<PackageWriter>("writer", pack, async_mode_, statistics_.pack_statistics);
@@ -885,9 +985,12 @@ void ArchiveSlice::destroy(td::Promise<td::Unit> promise) {
   before_query();
   destroyed_ = true;
 
-  for (auto &p : packages_) {
-    td::unlink(p.path).ensure();
+  if (mode_ == td::DbOpenMode::db_primary) {
+    for (auto &p : packages_) {
+      td::unlink(p.path).ensure();
+    }
   }
+  
   if (statistics_.pack_statistics) {
     statistics_.pack_statistics->record_close(packages_.size());
   }
@@ -895,12 +998,17 @@ void ArchiveSlice::destroy(td::Promise<td::Unit> promise) {
   id_to_package_.clear();
   kv_ = nullptr;
 
-  delay_action([name = db_path_, attempt = 0,
-                promise = std::move(promise)]() mutable { destroy_db(name, attempt, std::move(promise)); },
-               td::Timestamp::in(0.0));
+  if (mode_ == td::DbOpenMode::db_primary) {
+    delay_action([name = db_path_, attempt = 0,
+                  promise = std::move(promise)]() mutable { destroy_db(name, attempt, std::move(promise)); },
+                td::Timestamp::in(0.0));
+  } else {
+    promise.set_value(td::Unit());
+  }
 }
 
 BlockSeqno ArchiveSlice::max_masterchain_seqno() {
+  before_query();
   auto key = get_db_key_lt_desc(ShardIdFull{masterchainId});
   std::string value;
   auto F = kv_->get(key, value);
@@ -916,6 +1024,33 @@ BlockSeqno ArchiveSlice::max_masterchain_seqno() {
   }
   auto last_idx = g->last_idx_ - 1;
   auto db_key = get_db_key_lt_el(ShardIdFull{masterchainId}, last_idx);
+  F = kv_->get(db_key, value);
+  F.ensure();
+  CHECK(F.move_as_ok() == td::KeyValue::GetStatus::Ok);
+  auto E = fetch_tl_object<ton_api::db_lt_el_value>(td::BufferSlice{value}, true);
+  E.ensure();
+  auto e = E.move_as_ok();
+  return e->id_->seqno_;
+}
+
+BlockSeqno ArchiveSlice::min_masterchain_seqno() {
+  before_query();
+  auto key = get_db_key_lt_desc(ShardIdFull{masterchainId});
+  std::string value;
+  auto F = kv_->get(key, value);
+  F.ensure();
+  if (F.move_as_ok() == td::KeyValue::GetStatus::NotFound) {
+    return 0;
+  }
+  auto G = fetch_tl_object<ton_api::db_lt_desc_value>(value, true);
+  G.ensure();
+  auto g = G.move_as_ok();
+  if (g->first_idx_ == g->last_idx_) {
+    return 0;
+  }
+
+  auto first_idx = g->first_idx_;
+  auto db_key = get_db_key_lt_el(ShardIdFull{masterchainId}, first_idx);
   F = kv_->get(db_key, value);
   F.ensure();
   CHECK(F.move_as_ok() == td::KeyValue::GetStatus::Ok);
@@ -1029,6 +1164,7 @@ void ArchiveSlice::truncate_shard(BlockSeqno masterchain_seqno, ShardIdFull shar
 }
 
 void ArchiveSlice::truncate(BlockSeqno masterchain_seqno, ConstBlockHandle, td::Promise<td::Unit> promise) {
+  CHECK(mode_ == td::DbOpenMode::db_primary);
   if (temp_ || archive_id_ > masterchain_seqno) {
     destroy(std::move(promise));
     return;
